fix(db): widen i64-backed tasks/consolidation INT columns to BIGINT
The tasks priority + 6 counter columns and the 5 consolidation_run_metrics
count columns were created as Postgres INT (INT4) but their Rust structs
type them as i64. Non-macro query_as::<_, Task>(...) (list_ready, claim,
list_filtered) failed at runtime with 'Rust type i64 (INT8) is not
compatible with SQL type INT4', breaking the dispatcher and the Kanban
board; the matching macro queries only compiled against a stale .sqlx
cache.

Migration 25 ALTERs all 12 columns to BIGINT, aligning the schema to the
i64 Rust types and fixing macro + non-macro queries uniformly. Drops the
now-redundant i32 narrowing / int8 casts in the task writes/status and
consolidation INSERT paths, and regenerates the .sqlx offline cache (17
query files now record INT8).

Columns deliberately left as INT (Rust field is i32/u32): org_config.id,
note_embeddings.embedding_dim, note_embedding_meta.embedding_dim,
verification_results.{step_index,exit_code}, code_chunks.{start_line,end_line}.

+++ server/crates/djinn-db/src/repositories/task/status.rs
@@ -188,7 +188,7 @@ impl TaskRepository {
                     let conflict_meta_ref = effective_conflict_metadata.as_deref();
 
                     // Apply all side effects atomically.
-                    let reopen_inc_val: i32 = if apply.increment_reopen { 1 } else { 0 };
+                    let reopen_inc_val: i64 = if apply.increment_reopen { 1 } else { 0 };
                     sqlx::query!(
                         r#"UPDATE tasks SET
                             status = $1,
@@ -309,7 +309,7 @@ impl TaskRepository {
                         ""
                     };
                     // Only increment reopen_count on actual reopen transitions (not open->open)
-                    let reopen_inc: i32 = if status == "open" && from_task.status != "open" {
+                    let reopen_inc: i64 = if status == "open" && from_task.status != "open" {
                         1
                     } else {
                         0
@@ -384,7 +384,7 @@ impl TaskRepository {
                         task_select_where_id!(id).fetch_one(self.db.pool()).await?;
 
                     // Only increment reopen_count on actual reopen transitions (not open->open)
-                    let reopen_inc: i32 = if status == "open" && from_task.status != "open" {
+                    let reopen_inc: i64 = if status == "open" && from_task.status != "open" {
                         1
                     } else {
                         0
